Tons of stuff...should've committed earlier :(
Changed the number stored in each chunk for tone or long-skip to a full
byte rather than just a 4-bit number. I think that's the main change
made to the erl files.

Added screen.coffee which implements the co-ordinate generation (sort of
based on a quad tree? Wikipedia's description is slightly different than
what I'm doing, but it's roughly that.)

There are at least 3 steps necessary to chop up a PNG image into a
Voronoi image, but I haven't put them into the Rakefile yet. The idea is
to use co-ordinates.coffee to generate the screen, run extract_png.rb to
pull out the PNG into the generated screen and do some naïve compressing
and dump that into co-ords_img.js or something like that, and then run
voronoi.coffee (writing its output to an html file from the shell) to
send the whole thing through d3 to generate the svg file (this last step
takes a super long time :(, unfortunately.)