Re-wrote [:skip, count] mechanism
Instead of skipping bounded sets of zero tones, extract_png.rb now uses
the voronoi neighbors of the point to decide whether it is redundant. As
a result, the removal of zero tones doesn't mess up  edges like it did
before, *and* the operation is also applied to redundant non-zero tones.
The number of polygons in the example was reduced by roughly 3,000 in
the test example (from roughly 30,000 polygons).

Also, reduced the path stroke width. Still not sure what it should be.